change int
change int is long int

+++ 6320500603_3.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main()
 {
-    int N;
-    scanf("%d",&N);
-    printf("%d %d",N%3,N%11);
+    long int N;
+    scanf("%ld",&N);
+    printf("%ld %ld",N%3,N%11);
     return 0;
 }
